Use an internal structure instead of using just the same "inherited" zend object. Effectively this makes new zend objects with persistent ids a proxy to the persistent internal struct on the persistent list.
The persistent object used to be directly allocated and stored in the
persistent list. Thus the newly created zend object data replaced bits
from all previously created objects with the same persistent id. This
might not be a problem at run time, but it causes double frees at tear
down.